refactor: extract ensure file function
