+++ packages/jerni/src/dev-cli/ensureFileExists.ts
@@ -1,0 +1,14 @@
+import fs from "node:fs";
+import yaml from "yaml";
+
+export default function ensureFileExists(filePath: string) {
+  if (!fs.existsSync(filePath)) {
+    const content = {
+      checksum: "",
+      events: [],
+    };
+
+    const stringifiedContent = yaml.stringify(content);
+    fs.writeFileSync(filePath, stringifiedContent);
+  }
+}
